add roles, move sequelize to config
